% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AvgTimeBookings.R
\name{AvgTimeBookings}
\alias{AvgTimeBookings}
\title{Cumulative sum of time bookings over package/project levels
This function is applied to all project types expect consultings since their unique code is shared}
\usage{
AvgTimeBookings(DF)
}
\arguments{
\item{DF}{result from e.g. filterData}
}
\value{
dataframe
}
\description{
Cumulative sum of time bookings over package/project levels
This function is applied to all project types expect consultings since their unique code is shared
}
\examples{
All_Tabs <- getPFData()
Data <- extractData(All_Tabs)
Data <- filterData(Data, All_Tabs)
Data <- AvgTimeBookings(Data)
}
