% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildTable.R
\name{BuildTable}
\alias{BuildTable}
\title{Print a datatable object}
\usage{
BuildTable(
  DataTab,
  OldCols = c("ProdDate", "TimeSpent", "TimeBudget", "TimePercent", "Fraction"),
  NewCols = c("Productive Date", "Time Spent (HH:MM)", "Time Budget (HH:MM)",
    "Time Spent/Budget [\%]", "Fraction Spent/All [\%]")
)
}
\arguments{
\item{DataTab}{result from e.g. Calculations}

\item{OldCols}{Optional: vector of strings including DataTab columns names to replace}

\item{NewCols}{Optional: vector of strings for the replacement columns which has to be the same length as OldCols}
}
\description{
Print a datatable object
}
\examples{
All_Tabs <- getPFData()
Data <- extractData(All_Tabs)
Data <- filterData(Data, All_Tabs)
Data <- Calculations(Data)
Data <- BuildTable(Data)
}
