% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculations.R
\name{Calculations}
\alias{Calculations}
\title{Computing calculations for HourlyCosts, MoneySpent and TimePercent variables}
\usage{
Calculations(Data)
}
\arguments{
\item{Data}{result from e.g. filterData}
}
\value{
dataframe
}
\description{
Computing calculations for HourlyCosts, MoneySpent and TimePercent variables
}
\examples{
All_Tabs <- getPFData()
Data <- extractData(All_Tabs)
Data <- filterData(Data, All_Tabs)
Data <- Calculations(Data)
}
