% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTime.R
\name{ConvertTime}
\alias{ConvertTime}
\title{Convert time in minutes as integers to standard HH:MM}
\usage{
ConvertTime(Vector)
}
\arguments{
\item{DataTab}{vector of time in minutes formatted as integers}
}
\value{
vector
}
\description{
Convert time in minutes as integers to standard HH:MM
}
\examples{
TimeasHHMM <- ConvertTime(200)
TimeasHHMM
"3:20"
}
