% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractData.R
\name{extractData}
\alias{extractData}
\title{Extract and format ProjectFacts data from the getPFData function}
\usage{
extractData(All_Tabs)
}
\arguments{
\item{All_Tabs}{result from e.g. getPFData}
}
\value{
dataframe
}
\description{
Extract and format ProjectFacts data from the getPFData function
}
\examples{
All_Tabs <- getPFData()
Data <- extractData(All_Tabs)

}
