% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterNotes.R
\name{FilterNotes}
\alias{FilterNotes}
\title{Extracting text information encapsulated between html-like tags, e.g. Notes field}
\usage{
FilterNotes(Vector, BegCode, EndCode)
}
\arguments{
\item{Vector}{vector containing strings for which to extract information}

\item{BegCode}{vector containing strings with beginning tags, e.g. "<92.259156585>"}

\item{EndCode}{vector containing strings with ending tags, e.g. "</92.259156585>"}
}
\value{
vector
}
\description{
Extracting text information encapsulated between html-like tags, e.g. Notes field
}
\examples{
Teststring <- <92.259156585> Hello world </92.259156585>
Data <- FilterNotes(Vector = Teststring, BegCode = "<92.259156585>", EndCode = "</92.259156585>")

}
