% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggColorHue.R
\name{ggColorHue}
\alias{ggColorHue}
\title{Emulate ggplot2 color palette
Taken from : https://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette}
\usage{
ggColorHue(n)
}
\arguments{
\item{n}{integer indicating the number of equally spaced hues around the color wheel, i.e. ggplot2-like}
}
\value{
vector
}
\description{
Emulate ggplot2 color palette
Taken from : https://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette
}
\examples{
ggColorHue(3)
"#F8766D" "#00BA38" "#619CFF"

}
