% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkPlot.R
\name{NetworkPlot}
\alias{NetworkPlot}
\title{Network plot function - Build the network
Tutorials:
https://www.statworx.com/en/content-hub/blog/interactive-network-visualization-with-r
http://datastorm-open.github.io/visNetwork/shiny.html}
\usage{
NetworkPlot(nodes, edges, layout = "layout_on_sphere", physics = T)
}
\arguments{
\item{nodes}{dataframe containing nodes data, see visNetwork()}

\item{edges}{dataframe containing edges data, see visNetwork()}

\item{layout}{igraph layouts to compute coordinates, see visIgraphLayout()}

\item{physics}{activate/deactivate physics properties, see visIgraphLayout()}

\item{params}{optional parameters to manipulate physics properties, see visPhysics()}
}
\description{
Network plot function - Build the network
Tutorials:
https://www.statworx.com/en/content-hub/blog/interactive-network-visualization-with-r
http://datastorm-open.github.io/visNetwork/shiny.html
}
\examples{
All_Tabs <- getPFData()
}
