% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp.R
\name{runCTUNetworkApp}
\alias{runCTUNetworkApp}
\title{CTU Network app}
\usage{
runCTUNetworkApp()
}
\description{
The CTU Network shiny app provides a dynamic network visualization of Projectfacts data
specific to CTU Bern. Through filtering options, the user can fine-tune the
input data which will influence the network representation as well as the bar graphs and table.
}
\details{
The app is also available at \href{https://shiny.ctu.unibe.ch/ctunetwork/}{https://shiny.ctu.unibe.ch/ctunetwork/}.
}
\examples{
# launch the app
\dontrun{
runCTUNetworkApp()
}

}
