% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardDates.R
\name{StandardDates}
\alias{StandardDates}
\title{Standardizing dates to European format, i.e. DD-MM-YYYY}
\usage{
StandardDates(DateVec)
}
\arguments{
\item{DateVec}{character containing a date}
}
\value{
character
}
\description{
Standardizing dates to European format, i.e. DD-MM-YYYY
}
\examples{
StandardDates("2022-12-01")
"01-12-2022"

}
