% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themeShiny.R
\name{themeShiny}
\alias{themeShiny}
\title{Customized ggplot2 theme applied to homogenize the plots in Shiny App}
\usage{
themeShiny(titleSize = 14)
}
\arguments{
\item{titleSize}{Optional: ggplot2 plot.title size argument}
}
\value{
list containing ggplot2 theme settings
}
\description{
Customized ggplot2 theme applied to homogenize the plots in Shiny App
}
\examples{
CustomTheme = themeShiny(titleSize = 12)

}
