% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterData.R
\name{filterData}
\alias{filterData}
\title{Filter ProjectFacts data from the getPFData function}
\usage{
filterData(Data, All_Tabs)
}
\arguments{
\item{Data}{result from e.g. extractData}

\item{All_Tabs}{result from e.g. getPFData}
}
\value{
dataframe
}
\description{
Filter ProjectFacts data from the getPFData function
}
\examples{
All_Tabs <- getPFData()
Data <- extractData(All_Tabs)
Data <- filterData(Data, All_Tabs)
}
