% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkPlot.R
\name{NetworkPlot}
\alias{NetworkPlot}
\title{Network plot function - Build the network
Tutorials:
https://www.statworx.com/en/content-hub/blog/interactive-network-visualization-with-r
http://datastorm-open.github.io/visNetwork/shiny.html}
\usage{
NetworkPlot(
  nodes,
  edges,
  params = list(layout = "layout_on_sphere", physics = T, solver =
    "hierarchicalRepulsion", timestep = 0.5, windX = 0, windY = 0, hierarchicalRepulsion
    = list(nodeDistance = 90, centralGravity = 0, springLength = 850, springConstant =
    0.01, damping = 0.09, avoidOverlap = 0))
)
}
\arguments{
\item{nodes}{dataframe containing nodes data, see visNetwork()}

\item{edges}{dataframe containing edges data, see visNetwork()}

\item{params}{list of parameters to control graph physics behaviour, see visIgraphLayout() and visPhysics()}
}
\description{
Network plot function - Build the network
Tutorials:
https://www.statworx.com/en/content-hub/blog/interactive-network-visualization-with-r
http://datastorm-open.github.io/visNetwork/shiny.html
}
\examples{
All_Tabs <- getPFData()
}
